+++ electron/main.js
@@ -1,4 +1,4 @@
 // This file is only a trampoline to load the actual main process file
 // from src/main/index.ts
 
-require('../src/main/index.ts');
+require('../src/main/index.js');
