+++ package.json
@@ -23,13 +23,17 @@
     "@electron/remote": "^2.0.9",
     "@element-plus/icons-vue": "^2.3.1",
     "@iconify-json/fa-solid": "^1.2.1",
+    "@types/express": "^5.0.1",
+    "@types/koa": "^2.15.0",
+    "@types/koa-bodyparser": "^4.3.12",
     "@unocss/vite": "66.1.0-beta.10",
     "better-sqlite3": "^8.3.0",
     "electron-store": "^8.1.0",
     "electron-vite": "^3.1.0",
     "element-plus": "^2.9.7",
     "express": "^5.1.0",
-    "@types/express": "^5.0.1",
+    "koa": "^2.16.1",
+    "koa-bodyparser": "^4.4.1",
     "pinia": "^2.0.30",
     "unocss": "66.1.0-beta.10",
     "vue": "^3.2.47",
